Base validator class
