+++ openapi_core/validation/request/validators.py
@@ -22,17 +22,10 @@ from openapi_core.validation.request.datatypes import (
     RequestParameters, RequestValidationResult,
 )
 from openapi_core.validation.util import get_operation_pattern
+from openapi_core.validation.validators import BaseValidator
 
 
-class RequestValidator(object):
-
-    def __init__(
-            self, spec,
-            custom_formatters=None, custom_media_type_deserializers=None,
-    ):
-        self.spec = spec
-        self.custom_formatters = custom_formatters
-        self.custom_media_type_deserializers = custom_media_type_deserializers
+class RequestValidator(BaseValidator):
 
     def validate(self, request):
         try:
@@ -80,25 +73,6 @@ class RequestValidator(object):
 
         body, body_errors = self._get_body(request, operation)
         return RequestValidationResult(body_errors, body, None, None)
-
-    def _get_operation_pattern(self, request):
-        server = self.spec.get_server(request.full_url_pattern)
-
-        return get_operation_pattern(
-            server.default_url, request.full_url_pattern
-        )
-
-    def _find_path(self, request):
-        operation_pattern = self._get_operation_pattern(request)
-
-        path = self.spec[operation_pattern]
-        path_variables = {}
-        operation = self.spec.get_operation(operation_pattern, request.method)
-        servers = path.servers or operation.servers or self.spec.servers
-        server = servers[0]
-        server_variables = {}
-
-        return path, operation, server, path_variables, server_variables
 
     def _get_security(self, request, operation):
         security = operation.security or self.spec.security
@@ -222,15 +196,6 @@ class RequestValidator(object):
             raise MissingRequestBody(request)
         return request.body
 
-    def _deserialise_media_type(self, media_type, value):
-        from openapi_core.deserializing.media_types.factories import (
-            MediaTypeDeserializersFactory,
-        )
-        deserializers_factory = MediaTypeDeserializersFactory(
-            self.custom_media_type_deserializers)
-        deserializer = deserializers_factory.create(media_type)
-        return deserializer(value)
-
     def _deserialise_parameter(self, param, value):
         from openapi_core.deserializing.parameters.factories import (
             ParameterDeserializersFactory,
@@ -239,27 +204,7 @@ class RequestValidator(object):
         deserializer = deserializers_factory.create(param)
         return deserializer(value)
 
-    def _cast(self, param_or_media_type, value):
-        # return param_or_media_type.cast(value)
-        if not param_or_media_type.schema:
-            return value
-
-        from openapi_core.casting.schemas.factories import SchemaCastersFactory
-        casters_factory = SchemaCastersFactory()
-        caster = casters_factory.create(param_or_media_type.schema)
-        return caster(value)
-
     def _unmarshal(self, param_or_media_type, value):
-        if not param_or_media_type.schema:
-            return value
-
-        from openapi_core.unmarshalling.schemas.factories import (
-            SchemaUnmarshallersFactory,
+        return super(RequestValidator, self)._unmarshal(
+            param_or_media_type, value, context=UnmarshalContext.REQUEST,
         )
-        unmarshallers_factory = SchemaUnmarshallersFactory(
-            self.spec._resolver, self.custom_formatters,
-            context=UnmarshalContext.REQUEST,
-        )
-        unmarshaller = unmarshallers_factory.create(
-            param_or_media_type.schema)
-        return unmarshaller(value)
